backend/requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
coreapi==2.3.3
coreschema==0.0.4
cryptography==38.0.3
Django==4.1.3
django-coreapi==2.3.0
django-credit-cards==0.4.1
django-filter==22.1
django-location-field==2.1.0
django-phone-field==1.8.1
django-phonenumber-field==7.0.0
django-phonenumbers==1.0.1
django-recurrence==1.11.1
django-rest-framework==0.1.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-yasg==1.21.4
idna==3.4
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.1
packaging==21.3
phonenumbers==8.13.0
phonenumberslite==8.13.0
Pillow==9.3.0
pipe==2.0
pycparser==2.21
PyJWT==2.6.0
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.6
requests==2.28.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
six==1.16.0
sqlparse==0.4.3
uritemplate==4.1.1
urllib3==1.26.12
